requirements.txt
python-docx>=1.1.0
bibtexparser>=1.4.0
Pillow>=10.0.0
